replace fmod on linux with openal, merge linux and win32 into desktop
