+++ cocos/audio/desktop/AudioCache.cpp
@@ -27,15 +27,15 @@
 
 #include "platform/CCPlatformConfig.h"
 
-#if CC_TARGET_PLATFORM == CC_PLATFORM_WIN32
+#if CC_TARGET_PLATFORM == CC_PLATFORM_WIN32 || CC_TARGET_PLATFORM == CC_PLATFORM_LINUX
 
-#include "audio/win32/AudioCache.h"
+#include "audio/desktop/AudioCache.h"
 #include <thread>
 #include "base/CCDirector.h"
 #include "base/CCScheduler.h"
 
-#include "audio/win32/AudioDecoderManager.h"
-#include "audio/win32/AudioDecoder.h"
+#include "audio/desktop/AudioDecoderManager.h"
+#include "audio/desktop/AudioDecoder.h"
 
 #define VERY_VERY_VERBOSE_LOGGING
 #ifdef VERY_VERY_VERBOSE_LOGGING
