+++ cocos/audio/desktop/AudioDecoder.cpp
@@ -23,65 +23,70 @@
  THE SOFTWARE.
  ****************************************************************************/
 
-#include "audio/win32/AudioDecoderOgg.h"
-#include "audio/win32/AudioMacros.h"
+#include "audio/desktop/AudioDecoder.h"
+#include "audio/desktop/AudioMacros.h"
 #include "platform/CCFileUtils.h"
 
-#define LOG_TAG "AudioDecoderOgg"
+#define LOG_TAG "AudioDecoder"
 
 namespace cocos2d {
 
-    AudioDecoderOgg::AudioDecoderOgg()
+AudioDecoder::AudioDecoder()
+    : _isOpened(false)
+    , _totalFrames(0)
+    , _bytesPerFrame(0)
+    , _sampleRate(0)
+    , _channelCount(0)
+    {
+
+    }
+
+    AudioDecoder::~AudioDecoder()
     {
     }
 
-    AudioDecoderOgg::~AudioDecoderOgg()
+
+    bool AudioDecoder::isOpened() const
     {
-        close();
+        return _isOpened;
     }
 
-    bool AudioDecoderOgg::open(const char* path)
+    uint32_t AudioDecoder::readFixedFrames(uint32_t framesToRead, char* pcmBuf)
     {
-        std::string fullPath = FileUtils::getInstance()->fullPathForFilename(path);
-        if (0 == ov_fopen(FileUtils::getInstance()->getSuitableFOpen(fullPath).c_str(), &_vf))
+        uint32_t framesRead = 0;
+        uint32_t framesReadOnce = 0;
+        do
         {
-            // header
-            vorbis_info* vi = ov_info(&_vf, -1);
-            _sampleRate = static_cast<uint32_t>(vi->rate);
-            _channelCount = vi->channels;
-            _bytesPerFrame = vi->channels * sizeof(short);
-            _totalFrames = static_cast<uint32_t>(ov_pcm_total(&_vf, -1));
-            _isOpened = true;
-            return true;
-        }
-        return false;
-    }
+            framesReadOnce = read(framesToRead - framesRead, pcmBuf + framesRead * _bytesPerFrame);
+            framesRead += framesReadOnce;
+        } while (framesReadOnce != 0 && framesRead < framesToRead);
 
-    void AudioDecoderOgg::close()
-    {
-        if (isOpened())
+        if (framesRead < framesToRead)
         {
-            ov_clear(&_vf);
-            _isOpened = false;
+            memset(pcmBuf + framesRead * _bytesPerFrame, 0x00, (framesToRead - framesRead) * _bytesPerFrame);
         }
+
+        return framesRead;
     }
 
-    uint32_t AudioDecoderOgg::read(uint32_t framesToRead, char* pcmBuf)
+    uint32_t AudioDecoder::getTotalFrames() const
     {
-        int currentSection = 0;
-        int bytesToRead = framesToRead * _bytesPerFrame;
-        long bytesRead = ov_read(&_vf, pcmBuf, bytesToRead, 0, 2, 1, &currentSection);
-        return static_cast<uint32_t>(bytesRead / _bytesPerFrame);
+        return _totalFrames;
     }
 
-    bool AudioDecoderOgg::seek(uint32_t frameOffset)
+    uint32_t AudioDecoder::getBytesPerFrame() const
     {
-        return 0 == ov_pcm_seek(&_vf, frameOffset);
+        return _bytesPerFrame;
     }
 
-    uint32_t AudioDecoderOgg::tell() const
+    uint32_t AudioDecoder::getSampleRate() const
     {
-        return static_cast<uint32_t>(ov_pcm_tell(const_cast<OggVorbis_File*>(&_vf)));
+        return _sampleRate;
+    }
+
+    uint32_t AudioDecoder::getChannelCount() const
+    {
+        return _channelCount;
     }
 
 } // namespace cocos2d {
