+++ cocos/audio/desktop/AudioDecoderManager.h
