+++ cocos/audio/desktop/AudioEngine-desktop.h
@@ -1,5 +1,5 @@
 /****************************************************************************
- Copyright (c) 2015-2016 Chukong Technologies Inc.
+ Copyright (c) 2014-2016 Chukong Technologies Inc.
  Copyright (c) 2017-2018 Xiamen Yaji Software Co., Ltd.
 
  http://www.cocos2d-x.org
@@ -22,23 +22,17 @@
  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  THE SOFTWARE.
  ****************************************************************************/
-#include "platform/CCPlatformConfig.h"
+#pragma once
 
-#if CC_TARGET_PLATFORM == CC_PLATFORM_LINUX
-
-#ifndef __AUDIO_ENGINE_LINUX_H_
-#define __AUDIO_ENGINE_LINUX_H_
-
-#include <functional>
-#include <iostream>
-#include <map>
-#include "fmod.hpp"
-#include "fmod_errors.h"
-#include "audio/include/AudioEngine.h"
+#include <unordered_map>
 
 #include "base/CCRef.h"
+#include "audio/desktop/AudioCache.h"
+#include "audio/desktop/AudioPlayer.h"
 
 NS_CC_BEGIN
+
+class Scheduler;
 
 #define MAX_AUDIOINSTANCES 32
 
@@ -47,63 +41,44 @@ class CC_DLL AudioEngineImpl : public cocos2d::Ref
 public:
     AudioEngineImpl();
     ~AudioEngineImpl();
-    
+
     bool init();
     int play2d(const std::string &fileFullPath ,bool loop ,float volume);
     void setVolume(int audioID,float volume);
     void setLoop(int audioID, bool loop);
     bool pause(int audioID);
     bool resume(int audioID);
-    bool stop(int audioID);
+    void stop(int audioID);
     void stopAll();
     float getDuration(int audioID);
     float getCurrentTime(int audioID);
     bool setCurrentTime(int audioID, float time);
     void setFinishCallback(int audioID, const std::function<void (int, const std::string &)> &callback);
-    
+
     void uncache(const std::string& filePath);
     void uncacheAll();
-    
-
-    int preload(const std::string& filePath, std::function<void(bool isSuccess)> callback);
-    
+    AudioCache* preload(const std::string& filePath, std::function<void(bool)> callback);
     void update(float dt);
-    
-    /**
-     * used internally by ffmod callback 
-     */ 
-    void onSoundFinished(FMOD::Channel * channel); 
-    
-private:
-  
-    /**
-    * returns null if a sound with the given path is not found
-    */
-    FMOD::Sound * findSound(const std::string &path);
-  
-    FMOD::Channel * getChannel(FMOD::Sound *);
-  
-    struct ChannelInfo{
-        int id;
-        std::string path; 
-        FMOD::Sound * sound;
-        FMOD::Channel * channel; 
-        bool loop; 
-        float volume; 
-        std::function<void (int, const std::string &)> callback;
-    };
-    
-    std::map<int, ChannelInfo> mapChannelInfo;
 
-    std::map<std::string, int> mapId;
-    
-    std::map<std::string, FMOD::Sound *> mapSound;  
-    
-    FMOD::System* pSystem;
-    
+private:
+    void _play2d(AudioCache *cache, int audioID);
+
+    ALuint _alSources[MAX_AUDIOINSTANCES];
+
+    //source,used
+    std::unordered_map<ALuint, bool> _alSourceUsed;
+
+    //filePath,bufferInfo
+    std::unordered_map<std::string, AudioCache> _audioCaches;
+
+    //audioID,AudioInfo
+    std::unordered_map<int, AudioPlayer*>  _audioPlayers;
+    std::mutex _threadMutex;
+
+    bool _lazyInitLoop;
+
+    int _currentAudioID;
+    Scheduler* _scheduler;
 };
 
 NS_CC_END
-#endif // __AUDIO_ENGINE_LINUX_H_
-#endif
-
