+++ cocos/audio/desktop/AudioMacros.h
