+++ cocos/audio/desktop/AudioPlayer.h
