+++ templates/cocos2dx_files.json
@@ -2754,22 +2754,47 @@
         "external/json/stream.h", 
         "external/json/stringbuffer.h", 
         "external/json/writer.h", 
-        "external/linux-specific/fmod/CMakeLists.txt", 
-        "external/linux-specific/fmod/include/fmod.h", 
-        "external/linux-specific/fmod/include/fmod.hpp", 
-        "external/linux-specific/fmod/include/fmod_codec.h", 
-        "external/linux-specific/fmod/include/fmod_common.h", 
-        "external/linux-specific/fmod/include/fmod_dsp.h", 
-        "external/linux-specific/fmod/include/fmod_dsp_effects.h", 
-        "external/linux-specific/fmod/include/fmod_errors.h", 
-        "external/linux-specific/fmod/include/fmod_output.h", 
-        "external/linux-specific/fmod/prebuilt/64-bit/libfmod.so", 
-        "external/linux-specific/fmod/prebuilt/64-bit/libfmod.so.6", 
-        "external/linux-specific/fmod/prebuilt/64-bit/libfmodL.so", 
-        "external/linux-specific/fmod/prebuilt/64-bit/libfmodL.so.6", 
         "external/md5/CMakeLists.txt", 
         "external/md5/md5.c", 
         "external/md5/md5.h", 
+        "external/MP3Decoder/CMakeLists.txt", 
+        "external/MP3Decoder/include/mpg123.h", 
+        "external/MP3Decoder/prebuilt/linux/64-bit/libmpg123.a", 
+        "external/MP3Decoder/prebuilt/win32/libmpg123.dll", 
+        "external/MP3Decoder/prebuilt/win32/libmpg123.lib", 
+        "external/OggDecoder/CMakeLists.txt", 
+        "external/OggDecoder/include/ogg/ogg.h", 
+        "external/OggDecoder/include/ogg/os_types.h", 
+        "external/OggDecoder/include/vorbis/codec.h", 
+        "external/OggDecoder/include/vorbis/vorbisenc.h", 
+        "external/OggDecoder/include/vorbis/vorbisfile.h", 
+        "external/OggDecoder/prebuilt/linux/64-bit/libogg.a", 
+        "external/OggDecoder/prebuilt/linux/64-bit/libvorbis.a", 
+        "external/OggDecoder/prebuilt/linux/64-bit/libvorbisfile.a", 
+        "external/OggDecoder/prebuilt/win32/libogg.dll", 
+        "external/OggDecoder/prebuilt/win32/libogg.lib", 
+        "external/OggDecoder/prebuilt/win32/libvorbis.dll", 
+        "external/OggDecoder/prebuilt/win32/libvorbis.lib", 
+        "external/OggDecoder/prebuilt/win32/libvorbisfile.dll", 
+        "external/OggDecoder/prebuilt/win32/libvorbisfile.lib", 
+        "external/OpenalSoft/CMakeLists.txt", 
+        "external/OpenalSoft/include/linux/AL/al.h", 
+        "external/OpenalSoft/include/linux/AL/alc.h", 
+        "external/OpenalSoft/include/linux/AL/alext.h", 
+        "external/OpenalSoft/include/linux/AL/efx-creative.h", 
+        "external/OpenalSoft/include/linux/AL/efx-presets.h", 
+        "external/OpenalSoft/include/linux/AL/efx.h", 
+        "external/OpenalSoft/prebuilt/linux/64-bit/libopenal.so", 
+        "external/OpenalSoft/prebuilt/linux/64-bit/libopenal.so.1", 
+        "external/OpenalSoft/prebuilt/linux/64-bit/libopenal.so.1.17.0", 
+        "external/OpenalSoft/include/win32/AL/al.h", 
+        "external/OpenalSoft/include/win32/AL/alc.h", 
+        "external/OpenalSoft/include/win32/AL/alext.h", 
+        "external/OpenalSoft/include/win32/AL/efx-creative.h", 
+        "external/OpenalSoft/include/win32/AL/efx-presets.h", 
+        "external/OpenalSoft/include/win32/AL/efx.h", 
+        "external/OpenalSoft/prebuilt/win32/OpenAL32.dll", 
+        "external/OpenalSoft/prebuilt/win32/OpenAL32.lib", 
         "external/openssl/CMakeLists.txt", 
         "external/openssl/include/android/openssl/aes.h", 
         "external/openssl/include/android/openssl/asn1.h", 
@@ -3690,31 +3715,6 @@
         "external/win10-specific/zlib/prebuilt/x64/zlib.lib", 
         "external/win10-specific/zlib/prebuilt/x64/zlib1.dll", 
         "external/win10-specific/zlib/prebuilt/x64/zlibstatic.lib", 
-        "external/win32-specific/MP3Decoder/CMakeLists.txt", 
-        "external/win32-specific/MP3Decoder/include/mpg123.h", 
-        "external/win32-specific/MP3Decoder/prebuilt/libmpg123.dll", 
-        "external/win32-specific/MP3Decoder/prebuilt/libmpg123.lib", 
-        "external/win32-specific/OggDecoder/CMakeLists.txt", 
-        "external/win32-specific/OggDecoder/include/ogg/ogg.h", 
-        "external/win32-specific/OggDecoder/include/ogg/os_types.h", 
-        "external/win32-specific/OggDecoder/include/vorbis/codec.h", 
-        "external/win32-specific/OggDecoder/include/vorbis/vorbisenc.h", 
-        "external/win32-specific/OggDecoder/include/vorbis/vorbisfile.h", 
-        "external/win32-specific/OggDecoder/prebuilt/libogg.dll", 
-        "external/win32-specific/OggDecoder/prebuilt/libogg.lib", 
-        "external/win32-specific/OggDecoder/prebuilt/libvorbis.dll", 
-        "external/win32-specific/OggDecoder/prebuilt/libvorbis.lib", 
-        "external/win32-specific/OggDecoder/prebuilt/libvorbisfile.dll", 
-        "external/win32-specific/OggDecoder/prebuilt/libvorbisfile.lib", 
-        "external/win32-specific/OpenalSoft/CMakeLists.txt", 
-        "external/win32-specific/OpenalSoft/include/AL/al.h", 
-        "external/win32-specific/OpenalSoft/include/AL/alc.h", 
-        "external/win32-specific/OpenalSoft/include/AL/alext.h", 
-        "external/win32-specific/OpenalSoft/include/AL/efx-creative.h", 
-        "external/win32-specific/OpenalSoft/include/AL/efx-presets.h", 
-        "external/win32-specific/OpenalSoft/include/AL/efx.h", 
-        "external/win32-specific/OpenalSoft/prebuilt/OpenAL32.dll", 
-        "external/win32-specific/OpenalSoft/prebuilt/OpenAL32.lib", 
         "external/win32-specific/gles/CMakeLists.txt", 
         "external/win32-specific/gles/include/OGLES/GL/glew.h", 
         "external/win32-specific/gles/include/OGLES/GL/glxew.h",
